V++; CommandException accepts to-be-formatted input
Use-Case: If the program left the command scope and is somewhere, it can still throw this exception. At that point there is no Feedback class to use, formatting is impossible. Thats why it will be formatted on printing now.
Any plugin using this exception, has to remove the formatting now.